[Add & Update] 支払い方法のページを作成 #12
今回の実装に関連した部分で、Model/Userが間違っていたため
その部分だけ修正を行った。他にバグが含まれているかどうかはわからない。

+++ OKN/app/Http/Controllers/PaymentController.php
@@ -16,6 +16,7 @@ class PaymentController extends Controller
     public function index()
     {
         //
+        return view('payments.index', ["items" => Auth::user()->payments()->get()]);
     }
 
     /**
@@ -25,7 +26,7 @@ class PaymentController extends Controller
      */
     public function create()
     {
-        return view('payments.create');
+        return view('payments.create', ['paymentGenres' => Auth::user()->paymentGenres()->get()]);
     }
 
     /**
@@ -36,20 +37,13 @@ class PaymentController extends Controller
      */
     public function store(Request $request)
     {
-        try{
-            $payment = new Payment;
-            $user = Auth::user();
-            $payment->name = $request->name;
-            if($request->filled('paymentGenre_id'))
-                $payment->paymentGenre_id = $user->PaymentGenre()->findOrFail($request->paymentGenre_id)->id;
-            $payment->memo = $request->memo;
-            $user->payment()->save($payment);
-            if($request->filled('parent')) {
-                $parent = $user->payment()->findOrFail($request->parent);
-                $parent->child()->attach($payment->id);
-            }
-        }catch(Exception $e){}
-        return redirect('payments/'.$payment->id);
+        $payment = new Payment;
+        $user = Auth::user();
+        $payment->name = $request->name;
+        $payment->memo = $request->memo;
+        $payment->paymentGenre = $request->paymentGenre;
+        $user->payments()->save($payment);
+        return redirect()->route('payments.show', $payment->id);
     }
 
     /**
@@ -60,8 +54,8 @@ class PaymentController extends Controller
      */
     public function show(Payment $payment)
     {
-        //
-        return ["payment" => $payment];
+        if($payment->user_id != Auth::id()) return \App::abort(404);
+        return view('payments.show', ["item" => $payment]);
     }
 
     /**
@@ -72,7 +66,8 @@ class PaymentController extends Controller
      */
     public function edit(Payment $payment)
     {
-        return view('payments.edit');
+        if($payment->user_id != Auth::id()) return \App::abort(404);
+        return view('payments.edit', ["item" => $payment, "paymentGenres" => Auth::user()->paymentGenres()->get()]);
     }
 
     /**
@@ -84,20 +79,12 @@ class PaymentController extends Controller
      */
     public function update(Request $request, Payment $payment)
     {
-        //
-        try{
-            $payment = new Payment;
-            $payment->name = $request->name;
-            if($request->filled('paymentGenre_id'))
-                $payment->paymentGenre_id = $user->PaymentGenre()->findOrFail($request->paymentGenre_id)->id;
-            $payment->memo = $request->memo;
-            if($request->filled('parent')) {
-                $parent = $user->payment()->findOrFail($request->parent);
-                $parent->child()->attach($payment->id);
-            }
-            $payment->save();
-        }catch(Exception $e){}
-        return redirect('payments/'.$payment->id);
+        if($payment->user_id != Auth::id()) return \App::abort(404);
+        $payment->name = $request->name;
+        $payment->memo = $request->memo;
+        $payment->paymentGenre = $request->paymentGenre;
+        $payment->save();
+        return redirect()->route('payments.show', $payment->id);
     }
 
     /**
@@ -108,7 +95,8 @@ class PaymentController extends Controller
      */
     public function destroy(Payment $payment)
     {
-        //
+        if($payment->user_id != Auth::id()) return \App::abort(404);
         $payment->delete();
+        return redirect()->route('payments.index');
     }
 }
